web_app/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.10.0
python-engineio==4.8.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
eventlet==0.34.2
